Move fixture to own file
